+++ bindings/nostr-sdk-ffi/src/protocol/nips/mod.rs
@@ -5,6 +5,7 @@
 pub mod nip01;
 pub mod nip04;
 pub mod nip05;
+pub mod nip09;
 pub mod nip10;
 pub mod nip11;
 pub mod nip13;