+++ bindings/nostr-sdk-js/src/protocol/nips/mod.rs
@@ -6,6 +6,7 @@ pub mod nip01;
 pub mod nip04;
 pub mod nip05;
 pub mod nip07;
+pub mod nip09;
 pub mod nip11;
 pub mod nip15;
 pub mod nip19;