+++ crates/nostr/src/prelude.rs
@@ -38,6 +38,7 @@ pub use crate::nips::nip05::{self, *};
 pub use crate::nips::nip06::{self, *};
 #[cfg(all(feature = "nip07", target_arch = "wasm32"))]
 pub use crate::nips::nip07::{self, *};
+pub use crate::nips::nip09::{self, *};
 pub use crate::nips::nip10::{self, *};
 #[cfg(all(feature = "std", feature = "nip11"))]
 pub use crate::nips::nip11::{self, *};